+++ HISTORY.rst
@@ -21,6 +21,12 @@ TODO
 
 * displaying dates when ``print(Submission)`` instances
 * get a Team instance from Submission instance
+* The number of seconds is not human-readable, better to use hh:mm:ss
+* Set a default date for ``releaseDate`` (``str(datetime.now().date())`` could be
+  ok)
+* Filter a sample by status (ex. pending validation)
+* compiling PDF using SVG images
+* Using namespaces to configure API endpoints
 
 
 Features