+++ HISTORY.rst
@@ -21,11 +21,9 @@ TODO
 
 * displaying dates when ``print(Submission)`` instances
 * get a ``Team`` instance from ``Submission`` instance
-* The number of seconds is not human-readable, better to use ``hh:mm:ss``
 * Set a default date for ``releaseDate`` (``str(datetime.now().date())`` could be
   ok)
 * Filter a sample by status (ex. pending validation)
-* compiling PDF using PNG images (change badges)
 * fix this stuff::
 
     submission = team.get_submissions()[0]
@@ -40,6 +38,8 @@ TODO
 Features
 ^^^^^^^^
 
+* str(auth) will report duration in ``hh:mm:ss``
+* compiling PDF using PNG images (change badges)
 * raise no exceptions where no team is found (returning an empty list)
 * Using namespaces to configure API endpoints (``pyUSIrest.settings``)