+++ HISTORY.rst
@@ -10,11 +10,15 @@ TODO
 
 * Model custom exceptions
 * get a ``Team`` instance from ``Submission`` instance
-* Set a default date for ``releaseDate`` (``str(datetime.now().date())`` could be ok)
+* ``Submission.has_errors`` make two identical queries, on to determine the
+  status and one to search errors, simplify it by doing only a query
+* filtering sample by status or error make a lot of queries. Consider writing
+  coroutines or reading ValidationResult as pages
 
 Features
 ^^^^^^^^
 
+* Set a default date if ``releaseDate`` attribute is missing
 * improved documentation by describing how to sort and filter objects
 * fix bug when adding samples to a submission retrieved with ``team.get_submission()``
 * Update documentation. Set ``taxon`` in sample data (mandatory attribute)